bug fix: exception handling
The operand stack diagram for the athrow instruction may be misleading:
If a handler for this exception is matched in the current method, the
athrow instruction discards all the values on the operand stack, then
pushes the thrown object onto the operand stack. However, if no handler
is matched in the current method and the exception is thrown farther up
the method invocation chain, then the operand stack of the method (if
any) that handles the exception is cleared and objectref is pushed onto
that empty operand stack. All intervening frames from the method that
threw the exception up to, but not including, the method that handles
the exception are discarded.